:recycle: Refactor: 개인 관련 도메인 이동
<body>
- individual/domain : Alarm/Category/Image/Palette/Schedule class
- individual/dto : QueryDSL dto
- individual/repository : repository
- individual/type : enum 등의 type class

<footer>
- 관련: #141
