+++ application/external-api/src/main/java/com/namo/spring/application/external/api/user/facade/UserFacade.java
@@ -28,16 +28,16 @@ import com.namo.spring.application.external.domain.group.application.impl.MoimAn
 import com.namo.spring.application.external.domain.group.application.impl.MoimMemoLocationService;
 import com.namo.spring.application.external.domain.group.application.impl.MoimScheduleAndUserService;
 import com.namo.spring.application.external.domain.group.domain.MoimScheduleAndUser;
-import com.namo.spring.application.external.domain.individual.application.converter.CategoryConverter;
-import com.namo.spring.application.external.domain.individual.application.impl.AlarmService;
-import com.namo.spring.application.external.domain.individual.application.impl.CategoryService;
-import com.namo.spring.application.external.domain.individual.application.impl.ImageService;
-import com.namo.spring.application.external.domain.individual.application.impl.PaletteService;
-import com.namo.spring.application.external.domain.individual.application.impl.ScheduleService;
-import com.namo.spring.application.external.domain.individual.domain.Category;
-import com.namo.spring.application.external.domain.individual.domain.Image;
-import com.namo.spring.application.external.domain.individual.domain.Schedule;
-import com.namo.spring.application.external.domain.individual.domain.constant.CategoryKind;
+import com.namo.spring.application.external.api.individual.converter.CategoryConverter;
+import com.namo.spring.application.external.api.individual.service.AlarmService;
+import com.namo.spring.application.external.api.individual.service.CategoryService;
+import com.namo.spring.application.external.api.individual.service.ImageService;
+import com.namo.spring.application.external.api.individual.service.PaletteService;
+import com.namo.spring.application.external.api.individual.service.ScheduleService;
+import com.namo.spring.db.mysql.domains.individual.domain.Category;
+import com.namo.spring.db.mysql.domains.individual.domain.Image;
+import com.namo.spring.db.mysql.domains.individual.domain.Schedule;
+import com.namo.spring.db.mysql.domains.individual.type.CategoryKind;
 import com.namo.spring.application.external.api.user.converter.TermConverter;
 import com.namo.spring.db.mysql.domains.user.domain.Term;
 import com.namo.spring.db.mysql.domains.user.domain.User;
