+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/application/MoimScheduleFacade.java
@@ -25,12 +25,12 @@ import com.namo.spring.application.external.domain.group.domain.MoimScheduleAlar
 import com.namo.spring.application.external.domain.group.domain.MoimScheduleAndUser;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupScheduleRequest;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupScheduleResponse;
-import com.namo.spring.application.external.domain.individual.application.impl.CategoryService;
-import com.namo.spring.application.external.domain.individual.application.impl.ScheduleService;
-import com.namo.spring.application.external.domain.individual.domain.Category;
-import com.namo.spring.application.external.domain.individual.domain.Schedule;
-import com.namo.spring.application.external.domain.individual.domain.constant.Location;
-import com.namo.spring.application.external.domain.individual.domain.constant.Period;
+import com.namo.spring.application.external.api.individual.service.CategoryService;
+import com.namo.spring.application.external.api.individual.service.ScheduleService;
+import com.namo.spring.db.mysql.domains.individual.domain.Category;
+import com.namo.spring.db.mysql.domains.individual.domain.Schedule;
+import com.namo.spring.db.mysql.domains.individual.type.Location;
+import com.namo.spring.db.mysql.domains.individual.type.Period;
 import com.namo.spring.application.external.api.user.service.UserService;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.application.external.global.common.constant.FilePath;
