+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/MoimSchedule.java
@@ -20,8 +20,8 @@ import jakarta.persistence.Table;
 
 import com.namo.spring.application.external.domain.group.domain.constant.MoimScheduleStatus;
 
-import com.namo.spring.application.external.domain.individual.domain.constant.Location;
-import com.namo.spring.application.external.domain.individual.domain.constant.Period;
+import com.namo.spring.db.mysql.domains.individual.type.Location;
+import com.namo.spring.db.mysql.domains.individual.type.Period;
 
 import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
