--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/domain/constant/CategoryStatus.java
@@ -1,5 +1,0 @@
-package com.namo.spring.application.external.domain.individual.domain.constant;
-
-public enum CategoryStatus {
-	ACTIVE, DELETE
-}
