--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/repository/alarm/AlarmRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.namo.spring.application.external.domain.individual.repository.alarm;
-
-public interface AlarmRepositoryCustom {
-}
