+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/domain/Alarm.java
@@ -1,4 +1,4 @@
-package com.namo.spring.application.external.domain.individual.domain;
+package com.namo.spring.db.mysql.domains.individual.domain;
 
 import java.util.Objects;
