+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/domain/Category.java
@@ -1,5 +1,6 @@
-package com.namo.spring.application.external.domain.individual.domain;
+package com.namo.spring.db.mysql.domains.individual.domain;
 
+import com.namo.spring.db.mysql.domains.individual.type.CategoryStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -11,8 +12,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
-import com.namo.spring.application.external.domain.individual.domain.constant.CategoryKind;
-import com.namo.spring.application.external.domain.individual.domain.constant.CategoryStatus;
+import com.namo.spring.db.mysql.domains.individual.type.CategoryKind;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
