+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/domain/Schedule.java
@@ -1,9 +1,11 @@
-package com.namo.spring.application.external.domain.individual.domain;
+package com.namo.spring.db.mysql.domains.individual.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.namo.spring.db.mysql.domains.individual.type.Location;
+import com.namo.spring.db.mysql.domains.individual.type.Period;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
@@ -16,8 +18,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 
-import com.namo.spring.application.external.domain.individual.domain.constant.Location;
-import com.namo.spring.application.external.domain.individual.domain.constant.Period;
 import com.namo.spring.core.common.exception.IndividualException;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 
@@ -42,7 +42,7 @@ public class Schedule extends BaseTimeEntity {
 	private String name;
 
 	@Embedded
-	Period period;
+    Period period;
 
 	@Embedded
 	private Location location;
