+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/dto/DiaryResponse.java
@@ -1,0 +1,19 @@
+package com.namo.spring.db.mysql.domains.individual.dto;
+
+import com.namo.spring.db.mysql.domains.individual.domain.Image;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+public class DiaryResponse {
+    @AllArgsConstructor
+    @Getter
+    @Builder
+    public static class GetDiaryByUserDto {
+        private Long scheduleId;
+        private String contents;
+        private List<Image> images;
+    }
+}
