+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/repository/alarm/AlarmRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.namo.spring.db.mysql.domains.individual.repository.alarm;
+
+public interface AlarmRepositoryCustom {
+}
