+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/repository/category/CategoryRepositoryCustom.java
@@ -1,11 +1,11 @@
-package com.namo.spring.application.external.domain.individual.repository.category;
+package com.namo.spring.db.mysql.domains.individual.repository.category;
 
 import java.util.List;
 
 import org.springframework.data.repository.query.Param;
 
-import com.namo.spring.application.external.domain.individual.domain.Category;
-import com.namo.spring.application.external.domain.individual.domain.constant.CategoryKind;
+import com.namo.spring.db.mysql.domains.individual.domain.Category;
+import com.namo.spring.db.mysql.domains.individual.type.CategoryKind;
 
 import com.namo.spring.db.mysql.domains.user.domain.User;
