+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/repository/image/ImageRepositoryCustom.java
@@ -1,0 +1,7 @@
+package com.namo.spring.db.mysql.domains.individual.repository.image;
+
+import com.namo.spring.db.mysql.domains.individual.domain.Schedule;
+
+public interface ImageRepositoryCustom {
+	void deleteDiaryImages(Schedule schedule);
+}
