+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/individual/type/CategoryStatus.java
@@ -1,0 +1,5 @@
+package com.namo.spring.db.mysql.domains.individual.type;
+
+public enum CategoryStatus {
+	ACTIVE, DELETE
+}
